+++ lib/src/text_input_dialog/ios_text_input_dialog.dart
@@ -83,7 +83,7 @@ class _IOSTextInputDialogState extends State<IOSTextInputDialog> {
     }
 
     void cancel() => navigator.pop();
-    BoxDecoration _borderDecoration({
+    BoxDecoration borderDecoration({
       required bool isTopRounded,
       required bool isBottomRounded,
     }) {
@@ -143,7 +143,7 @@ class _IOSTextInputDialogState extends State<IOSTextInputDialog> {
                   autocorrect: field.autocorrect,
                   prefix: prefixText == null ? null : Text(prefixText),
                   suffix: suffixText == null ? null : Text(suffixText),
-                  decoration: _borderDecoration(
+                  decoration: borderDecoration(
                     isTopRounded: i == 0,
                     isBottomRounded: i == _textControllers.length - 1,
                   ),